progress abstracting Config tests
