+++ src/test/java/org/betonquest/betonquest/api/bukkit/config/custom/UnmodifiableConfigurationSectionTest.java
@@ -2,19 +2,25 @@ package org.betonquest.betonquest.api.bukkit.config.custom;
 
 import org.betonquest.betonquest.api.bukkit.config.util.AbstractConfigurationSectionTest;
 import org.bukkit.configuration.ConfigurationSection;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
-import org.junit.jupiter.api.parallel.Execution;
-import org.junit.jupiter.api.parallel.ExecutionMode;
+
+import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-@Execution(ExecutionMode.CONCURRENT)
-@SuppressWarnings("PMD.JUnitAssertionsShouldIncludeMessage")
 /**
  * This is a test for {@link UnmodifiableConfigurationSection}.
  */
+@SuppressWarnings("PMD.JUnitAssertionsShouldIncludeMessage")
 public class UnmodifiableConfigurationSectionTest extends AbstractConfigurationSectionTest {
+    /**
+     * The values in the configuration before the test did run.
+     */
+    private Map<String, Object> values;
+
     /**
      * Empty constructor
      */
@@ -27,7 +33,19 @@ public class UnmodifiableConfigurationSectionTest extends AbstractConfigurationS
         return new UnmodifiableConfigurationSection(super.getConfig());
     }
 
-    private void assertThrowsException(final Executable executable) {
+
+    @BeforeEach
+    public void beforeEach() {
+        super.beforeEach();
+        values = config.getValues(true);
+    }
+
+    @AfterEach
+    public void afterEach() {
+        assertEquals(values, config.getValues(true));
+    }
+
+    private void assertThrowsUnmodifiableException(final Executable executable) {
         final Exception exception = assertThrows(UnsupportedOperationException.class, executable);
         assertEquals("This config is unmodifiable", exception.getMessage());
     }
@@ -35,48 +53,49 @@ public class UnmodifiableConfigurationSectionTest extends AbstractConfigurationS
     @Test
     @Override
     public void testCreateSectionWithValuesOnExistingConfigPath() {
-        assertThrowsException(super::testCreateSectionWithValuesOnExistingConfigPath);
+        assertThrowsUnmodifiableException(super::testCreateSectionWithValuesOnExistingConfigPath);
     }
 
     @Test
     @Override
     public void testAddDefaultOnChildSection() {
-        assertThrowsException(super::testAddDefaultOnChildSection);
+        assertThrowsUnmodifiableException(super::testAddDefaultOnChildSection);
     }
 
     @Test
     @Override
     public void testSet() {
-        assertThrowsException(super::testSet);
+        assertThrowsUnmodifiableException(super::testSet);
     }
 
     @Test
     @Override
     public void testSetOnExistingConfigPath() {
-        assertThrowsException(super::testSetOnExistingConfigPath);
+        assertThrowsUnmodifiableException(super::testSetOnExistingConfigPath);
     }
 
     @Test
     @Override
     public void testCreateSectionOnExistingConfigPath() {
-        assertThrowsException(super::testCreateSectionOnExistingConfigPath);
+        assertThrowsUnmodifiableException(super::testCreateSectionOnExistingConfigPath);
     }
 
     @Test
     @Override
     public void testCreateSection() {
-        assertThrowsException(super::testCreateSection);
+        assertThrowsUnmodifiableException(super::testCreateSection);
     }
 
     @Test
     @Override
     public void testAddDefaultOnChildSectionOnExistingConfigPath() {
-        assertThrowsException(super::testAddDefaultOnChildSectionOnExistingConfigPath);
+        assertThrowsUnmodifiableException(super::testAddDefaultOnChildSectionOnExistingConfigPath);
     }
 
     @Test
     @Override
     public void testCreateSectionWithValues() {
-        assertThrowsException(super::testCreateSectionWithValues);
+        assertThrowsUnmodifiableException(super::testCreateSectionWithValues);
     }
 }
+
