+++ src/test/java/org/betonquest/betonquest/api/bukkit/config/util/ConfigurationTestInterface.java
@@ -1,7 +1,7 @@
 package org.betonquest.betonquest.api.bukkit.config.util;
 
 @SuppressWarnings({"unused", "PMD.CommentRequired", "PMD.JUnit4TestShouldUseTestAnnotation", "PMD.TooManyMethods"})
-public interface ConfigurationTestInterface extends ConfigurationSectionTestInterface {
+public interface ConfigurationTestInterface {
     void testAddDefaultOnRootSection();
 
     void testAddDefaultOnRootSectionOnExistingConfigPath();
